LayoutTests/imported/w3c/web-platform-tests/streams/piping/error-propagation-forward.any-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error2: error2!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!

FAIL Errors must be propagated forward: starts errored; preventAbort = false; fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = false; rejected abort promise |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = undefined (falsy); fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = null (falsy); fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = false (falsy); fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = 0 (falsy); fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = -0 (falsy); fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = NaN (falsy); fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort =  (falsy); fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = true (truthy) |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = a (truthy) |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = 1 (truthy) |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = Symbol() (truthy) |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = [object Object] (truthy) |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = true, preventCancel = true |this| is not a Promise
FAIL Errors must be propagated forward: starts errored; preventAbort = true, preventCancel = true, preventClose = true |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored while empty; preventAbort = false; fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored while empty; preventAbort = false; rejected abort promise |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored while empty; preventAbort = true |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored while empty; dest never desires chunks; preventAbort = false; fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored while empty; dest never desires chunks; preventAbort = false; rejected abort promise |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored while empty; dest never desires chunks; preventAbort = true |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored after one chunk; preventAbort = false; fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored after one chunk; preventAbort = false; rejected abort promise |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored after one chunk; preventAbort = true |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored after one chunk; dest never desires chunks; preventAbort = false; fulfilled abort promise |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored after one chunk; dest never desires chunks; preventAbort = false; rejected abort promise |this| is not a Promise
FAIL Errors must be propagated forward: becomes errored after one chunk; dest never desires chunks; preventAbort = true |this| is not a Promise
FAIL Errors must be propagated forward: shutdown must not occur until the final write completes |this| is not a Promise
FAIL Errors must be propagated forward: shutdown must not occur until the final write completes; preventAbort = true |this| is not a Promise
FAIL Errors must be propagated forward: shutdown must not occur until the final write completes; becomes errored after first write |this| is not a Promise
FAIL Errors must be propagated forward: shutdown must not occur until the final write completes; becomes errored after first write; preventAbort = true |this| is not a Promise
